% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_wpgpDatasets.R
\docType{data}
\name{wpgpDatasets}
\alias{wpgpDatasets}
\title{Here is the title of the dataset}
\format{data.frame}
\source{
This is a link to 
\href{ftp://ftp.worldpop.org.uk/assets/wpgpDatasets.csv}{the origin data}
}
\usage{
data("wpgpDatasets")
}
\description{
More description of the dataset is here
}
\keyword{datasets}
