% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpDownloadR.R
\name{wpgpGetAvalZonalStats}
\alias{wpgpGetAvalZonalStats}
\title{wpgpGetAvalZonalStats function will return a List of ZonalStats}
\usage{
wpgpGetAvalZonalStats(ISO3 = NULL, stat = "mean", quiet = TRUE,
  method = "auto")
}
\arguments{
\item{ISO3}{a 3-character country code}

\item{stat}{Either as character: 'mean', 'sum'.}

\item{quiet}{Download Without any messages if TRUE}

\item{method}{Method to be used for downloading files. Current download methods
are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto"}
}
\value{
list
}
\description{
wpgpGetAvalZonalStats function will return a List of ZonalStats
}
\examples{
wpgpGetAvalZonalStats(ISO3="ABW")
}
