% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpDownloadR.R
\name{wpgpListCountries}
\alias{wpgpListCountries}
\title{wpgpListCountries function will return a list of the country
avalible to download}
\usage{
wpgpListCountries(quiet = TRUE)
}
\arguments{
\item{quiet}{quiet If TRUE, suppress status messages (if any)}
}
\value{
Dataframe
}
\description{
wpgpListCountries function will return a list of the country
avalible to download
}
