% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpgpDownloadR.R
\name{wpgpListCountryDatasets}
\alias{wpgpListCountryDatasets}
\title{wpgpListCountryDatasets function will return a data frame of
avalible covariates for a country}
\usage{
wpgpListCountryDatasets(ISO3 = NULL, quiet = TRUE)
}
\arguments{
\item{ISO3}{a 3-character country code or vector of country codes}

\item{quiet}{Download Without any messages if TRUE}
}
\value{
Dataframe
}
\description{
wpgpListCountryDatasets function will return a data frame of
avalible covariates for a country
}
\examples{
wpgpListCountryDatasets(ISO3="NPL")

wpgpListCountryDatasets(ISO3=c("NPL","BTN") )
}
